add object id conversion to relation resolvers
