+++ packages/crud/server/index.ts
@@ -3,3 +3,5 @@ export * from "../index";
 export * from "./connector";
 
 export * from "./slug";
+
+export * from "./objectIdToString";
