+++ packages/mongo/connector.ts
@@ -3,7 +3,7 @@ import { VulcanModel } from "@vulcanjs/model";
 import { filterFunction } from "./mongoParams";
 
 import mongoose, { QueryOptions, FilterQuery } from "mongoose";
-import { Connector } from "@vulcanjs/crud/server";
+import { Connector, convertIdAndTransformToJSON } from "@vulcanjs/crud/server";
 import { debugVulcan } from "@vulcanjs/utils";
 const debugMongoose = debugVulcan("mongoose");
 
@@ -13,59 +13,6 @@ export type MongooseConnector<TModel = any> = Connector<
   QueryOptions,
   mongoose.Model<any, any>
 >;
-
-/**
- * Converts Mongo ObjectId to string
- *
- * Connectors are expected to use string ids
- * It was the default behaviour in Meteor for Mongo,
- * but Mongoose/raw Mongo default behaviour is to use ObjectId
- *
- * => we prefer string ids in Vulcan for a consistent representation, in particular
- * between the GraphQL client (that will always use string ids) and the server
- *
- * TODO: not sure why we need to turn the document into JSON though
- */
-type MongoDoc<TModel> = {
-  toJSON: () => TModel;
-  _id: { toString: () => string };
-};
-function convertIdAndTransformToJSON<TModel>(doc: MongoDoc<TModel>): TModel;
-function convertIdAndTransformToJSON<TModel>(
-  docs: Array<MongoDoc<TModel>>
-): Array<TModel>;
-function convertIdAndTransformToJSON<TModel>(
-  docOrDocs: MongoDoc<TModel> | Array<MongoDoc<TModel>>
-): TModel | Array<TModel> {
-  if (!Array.isArray(docOrDocs)) {
-    const doc = docOrDocs;
-    if (!doc)
-      throw new Error(
-        "Document is not defined during id transformation. You may have malformed documents in your database."
-      );
-    if (!doc._id) {
-      throw new Error(
-        `Document has no valid _id ${JSON.stringify(
-          document
-        )} during id transformation. You may use malformed document _id
-          in your database (coming from Meteor?)`
-      );
-    }
-    return { ...docOrDocs.toJSON(), _id: docOrDocs._id.toString() };
-  } else {
-    return docOrDocs.map((document) => {
-      if (!document._id) {
-        throw new Error(
-          `Document has no valid _id ${JSON.stringify(
-            document
-          )} during id transformation. You may use malformed document _id
-          in your database (coming from Meteor?)`
-        );
-      }
-      return { ...document.toJSON(), _id: document._id.toString() };
-    });
-  }
-}
 
 export interface MongooseConnectorOptions {
   /** Force a mongoose model. Use if you need to customize the schema or options */
